Fixed texlive.profile to use minimal instead of custom scheme
I don't really need ALL of the items in the collection. So I'm starting
with a minimal scheme and using tlmgr to add additional packages as needed.
The nice thing about this is anyone can now pull this container and use
another dockerfile to add additional packages using tlmgr according to
their needs.

Signed-off-by: Charo Nuguid <me@thegeekettespeaketh.com>